refactor(types): use `any` instead of `unknown`
Example code was unable to extend the base class because of weird
TypeScript errors.

There is a semantic difference between `any` and `unknown`. Using the
latter seems to cause more pain as TypeScript complains about some
restrictions. These errors reduce the ability to extend the interface
or the base class.

So, for best compatibility, `any` is used in more cases.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>