feat(vue): implement base class
The base class is vital for implementing the basic Vue interface.
The base class does not require sub classes to pass any data to a
constructor. There is no constructor. Previous versions of Vue 2
just simulated the proper usage of constructors. Hence, no assumptions
are made to maximise compatibility.

The required reference to the internal Vue component instance is being
set at runtime by a factory function, which will be passed to
`defineComponent` as `setup()` function.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>