feat(vue): implement base class constructor
The base class constructor gets hold on the internal Vue component
instance and reads all properties from it. These properties are
applied to the newly created instance.

Because this is done within the base constructor, all properties
defined by Vue are immediately available to all constructors and
methods in the the any base class. Nice!

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>