feat: traverse prototypes to find functions
The component decorator needs to create data and a setup function
directly on the class (a.k.a prototype). When calling the setup
function it must call all setup functions from all parent classes.

The implemented utility provides helper functions to collect all the
functions, directly set to the prototype.

In TypeScript class notation, these are static functions on the class.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>