feat: introduce life-cycle function mapper
The life-cycle function names of the legacy Options API differs from
the new Composition API. So, a mapper is introduced. Fortunately, the
function definitions seem to be backwards compatible.

The mapping object uses the old names as property names and defines
each as a function to call the new Composition API functions.

In order to keep the Composition API functions mockable, the
properties of the mapper are not directly assigned from the target
functions. Instead, wrapper functions are used to call the target
functions at runtime. Then any mocked or stubbed Composition API
function will still be used and tested.

Although this might seem a slow down at runtime, I am very confident
that JIT compilers will optimise and that there will be no penalty.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>