feat: map "computed" option API to composition API
A new utility function created proper composition API calls from the
`computed` component options.

In order to make them reactive, the whole Vue instance is wrapped
with a reactive Proxy wrapper. If a (computed) property is set via the
wrapper, all dependent other properties will be updated. This way,
the data values do not need to be made reactive explicitly.
In previous vue-class-component, all value properties are added as
mixins of `data`. Since mixins are not supported by Vue anymore, a
different mechanism is used now.

This utility will be used by the component decorator to setup the
newly created component.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>