refactor(legacy): `this` type of life-cycle hooks
The `this` type of the life-cycle-hooks is defined.

In Vue 2 and Vue 3, `this` context for the hook functions are set to
the Vue internal component. Vue 2 reused the "external" class
component internally. Hence, it contained all defined methods and
properties as defined for the component. In Vue 3, the internal
component is disassociated from the external class component. Because
this library defines a base class to mimic the Vue 2 internal
component and all class component inherit from this base class, the
`this` context is bound to the created class component instance.

There is one exception to this rule: the hook `beforeCreate` does not
have any class component instance available. Hence, its `this` context
is set to the Vue 3 internal instance.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>