fix(init): use of reactive wrapper
Computed values need to be based on reactive data storages. Otherwise
they will never detect any change and cache the initial value forever.
This is "not a bug" but an "intended feature" as stated by a member
of the core dev team of VueJS.

Therefore, the setup of the Cue instance must operate on a reactive
wrapper and on the original instance simultaneously in order to notify
all computed values of changes on values they read.

see:
* https://github.com/vuejs/core/issues/6130

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>