feat(component): introduce builder for setup
Initialising and setting up the component to be Vue compatible is a
cumbersome task. A lot of patching and making properties reactive is
involved.

Hence, a builder is introduced to make the setup easier and more
readable.

The builder will be able to track all changes and remove duplicate
patches. This will help keep performance impact to a minimum.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>