feat: map watcher definitions to composition API
The legacy definitions for watchers on properties are converted to
calls to the composition API.

Unlike the legacy counterparts, the Vue 3 watchers installed are
based on getters to read the watched value from a reactive version of
the component instance. Thus, a watcher will be notified upon changes
to the watched property on the reactive instance.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>