refactor(builder): remove reactive property handling
There is no need to define reactive properties anymore. Reactivity
is achieved by using a reactive wrapper on the whole component
instance.

Handling reactive properties brought-in the problem of redefining
already reactive properties where some code holds a reference to
the old reactive property. No solution exists to this problem yet.
So, it is better to remove this altogether than run into problems
if someone is using this. There should be no necessity to make any
property reactive if just the reactive instance-wrapper is
used instead.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>