feat(builder): init watchers at the end of setup
Watchers are notified of changes on properties. If they are marked
with option `immediate`, then they will be executed immediately
upon creation. Unfortunately, the instance may not be fully
initialised yet or the watched value of the property not yet set.
Hence, the watchers are created only at the very end of the setup,
when `build()` is called from the builder. This will make sure, that
the whole instance has properly been setup and all watched values are
available.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>