reafactor(builder): params to helper func optional
THe parameters to the builder helper functions are optional now.
The helper functions check the validity anyway. Since the related
properties in the component option data are optional too, this change
makes life easier and code more readable because no separate if-clause
checks need to be applied prior to calling the functions.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>