fix(vue): `VueComponentBaseImpl` instance without Vue
For testing purposes, it is necessary to create an instance of the
`Vue` base class `VueComponentBaseImpl` without the availability of
any Vue internal instance. So, `getCurrentInstannce()` would return
`undefined`.

Previously, an error was thrown, which made it impossible to write
unit tests without creating a Vue component.
Now, this silently accepted and consumers of the Vue internal instance
`this.$` need to take care for themselves.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>