feat(builder): let builder create instance
The builder will be used to create the component instance. This way
the builder has access to the component class and thus access to all
options.

Using a builder instance makes testing much easier as it is easy to
mock/spy methods on an object but hard to mock global functions.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>