feat(builder): use component class with builder
Instead of providing an already created instance, a class for the
component is provided to the builder. Now, the builder is able to
read the original component options and call `beforeCreate()` hook
functions prior to creating the instance.

Using a builder instance enables tests to mock the builder in parts
and this test the functions better. This would not be as easy using
global functions instead.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>