refactor(builder): methods return builder class
For "Fluent API" the builder methods return the builder class itself
instead of the interface. The interface does not contain all the
functions the builder class provides. The interface represents the
builder API for custom setup functions whereas the builder
implementation class is used internally and has more features for
the proposed component decorators.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>