feat(builder): get component options from class
The component decorator stores the original component options with
the class. The builder provides a function to retrieve these options
from the class.

In order to support mixins of component classes, the options are
stored as a static filed to the constructor (class) and as a dynamic
function to the instance. So, in case component classes have been
mixed-in, stripping the constructor options, an instance is created
instead and the options retrieved from this instance.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>