fix(vue): _applyProperties use `toRaw(this)`
When creating new properties these should be created on the raw
instance and not on the reactive proxy, although the reactive proxy
will pass new properties to the raw instance. Nevertheless, this
behaviour might change in future versions of Vue or might misbehave.
Hence, the raw instance is retrieved from the reactive instance and
the properties to apply are created on the raw instance.
This will avoid future side effects in case Vue reactive proxy changes
its implementation or behaviour.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>