feat(decorator): implement `Component` decorator
The implementation of the component decorator creates a special
property `__vccOpts` on the class constructor function. This property
is the marker for Vue to use that content of that property to define
the component.

Lukyly, the most basic component only needs a list of properties and
a `setup` function to create the "external" component instance. This
factory function is used to create and prepare the Vue 3 component.
All defined options are read at runtime to allow mixins to alter the
class hierarchy easily.

Unfortunately, the SFC compiler strips the class constructor function.
Hence, at the moment there is no way base your child class on a
component exported inside a SFC file. TypeScript does not not know the
exported type of a SFC file anyways.

Properties of newly created instances are initialised as soon as
possible, in the Base class constructor. Nevertheless, all child
constructors may try to apply any initialisation value. Therefore, a
special setter for each property is needed to simply ignore the new
value. Otherwise, initialisation values in the class would overwrite
values passed-doen from Vue.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>