chore(build): include dist/* for release via git
Include the files in dist/* to enable using this package as
dependency directly loaded from the git repository.

As long as this package is not published to any NPM repository, using
git is a convenient way to include it. This needs to be supported by
adding the release files to the repository.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>