feat(utils): proxy to read missing from alternative
Function `createProxyRedirectReads` is introduced, creating a proxy
object to read missing properties from an alternative object.

This proxy will be used to provide all methods of the component to
the decorators but without really applying them to the options data.
All methods in the options data need to be applied to the component
as these options might alter the original ones. Because there is no
need to re-apply all already defined methods, original methods are to
be ignored. Therefore, only altered methods need to be applied to
the component. The proxy will help to filter for altered methods and
ignore all unaltered ones.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>