feat(decorator): provide methods to decorators
Decorators that need to alter the methods of the component need access
to the method to change. The previous component decorator provided
all methods via options. Thus, this component decorator need to do the
same and provide all methods via `options` data.

However, a proxy is used to filter for newly created or altered
functions and ignore all unchanged when applying these new definitions
to the component.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>