fix(build): rename `preinstall` script
All `preinstall` script of any dependencies are executed by NPM.
Hence, if this package is used as a dependency, its script gets
executed, too. Unfortunately, the git configuration is only meant to
be in place for its own repository and not for dependants. Colliding
with configs of dependants are possible and likely.
Therefore, one must be extremely careful with defining such a script
in a package that might be used by other packages.

For this reason, the git configuration will not be an automatic step
but must be applied manually from now on.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>