feat(setup): provide SetupContext to component
The setup context may contain valid data for the render function.
Hence, the context is passed to the newly created component and stored
as its property for later use.

Passing the context is performed via global storage, much like passing
the current Vue internal component instance to the external instance.
One can dispute the usage of global variables but this seems the
direct way forward and is already used with the Vue instance.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>