feat(render): call custom render hook
If the component is not defined in a Single File Component (SFC), then
the default render function is used to call the render hook function.
Compiling SFCs will overwrite the default render function with the
compiled SFC template. This is intended! There is no way to determine,
how to "merge" both functions. So, SFC templates override custom
render hooks!

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>