fix(vue): dummy render func use real arguments
The arguments passed to the render function are different than
expected. The passed argument is not the instance of the Vue
component class but some proxy object, containing only the
properties. All the functions are missing.

Fortunately, the real instance - with all the functions - is
available from undocumented property `setupState`. It is extracted
and the render function of the instance called - if available.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>