chore(build): disable minify
Minifying the source should only be done with the app building. The
app will minify if it is required.
There is no need for a library to minify or uglify, as the app build
process will do anyway.

Clear-text code will help a great deal in debugging your code.
Hence, the library is NOT minified.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>