feat(component): custom render func calls parent
If the TypeScript class based Vue component is not defined in a
"Single File Component" (SFC), no render function is created
automatically. Then the custom render function of the component
decorator is applied instead.

However, the base class might have a render function. Hence, if the
instance does not have a render hook function of its own, the render
function from the base class is used instead. The base class might
be defined in an SFC to leverage the Vue 3 template compiler and
its capabilities.

Unlike the original component decorator of Vue 2, this enables
components to be descendants of other SFC defined components and still
have proper rendering functions.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>