refactor(utils): add child class to base class list
Some functions need to include the provided class to the list of
"base" classes to form a full chain of all classes from current child
class to top parent class. Previously, those consumers pushed the
current child class to the end of the list on their own.
Because this use case if used often, an additional parameter makes the
function `getAllBaseClasses()` to do this for the caller.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>