fix(builder): build patched context for rendering
The builder is used to build the return value of `__vccOpts.setup()`.
Its return value is used by Vue as `this`-context to the rendering
function - but not unchanged.

Vue wraps the returned object with a proxy that checks every property
to be in the list of "own properties". Hence, all inherited properties
are ignored and lead to an error in the rendered template. Inherited
properties are NOT directly accessible to the template because of
this test.

The patched instance returned by the builder is a proxy itself that
returns `true` for all inherited properties as well. So, Vue is
tricked in believing the property is part of the list of
"own properties" and passes the call to class instance.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>