fix(builder): bind `this` context for render instance
The object returned by `setup()` is used by Vue to create the render
context for the template. Vue does not expect it to be a class.
On contrary, all properties are provided to the template as-is without
passing-in any `this` context to functions - even with event handlers.

In order to make the event handler work as expected, the methods are
provided with `this` context bound to the reactive instance of the
class instance.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>