fix(decorator): scope decorator to constructor
If a decorator function is added to the list of decorators in the
property `__decoratpr__` on the constructor function, the
prototype chain was followed to the property in the parent class.
This is bad as it does not separate parent class from sub classes and
effectively blocks any component inheritance.

This commit narrows the scope of `__decorators__` property to the
very specific constructor function. The property must exist on the
own property list and is created there if it does not exist.

Now, multiple children may be derived from the parent component
without interfering each other. Before this change, properties added
by a child have been defined on all other children as well, which is
totally unexpected and utterly wrong!

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>