fix(vue)!: make Vue type compatible with Vue 3
Previously, the type `Vue` was totally separate definition from Vue 3,
based on the definition of Vue 2. This turned out to be a problem with
templates as such type was unable to be used with dynamic components
and the tag `<component :is=... />`. Now, Vue type definition is
directly derived from Vue 3 export. All usage of this type are made
compatible to Vue 2, so no this should not break.
Nevertheless, chances are that this change will break legacy Vue 2
applications.

BREAKING CHANGE: type of Vue has changed!

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>