refactor(builder)!: init component builder optional
The initialisation of the component builder with a ready instance is
optional now. Instead, a class can be provided to create an instance
later when needed.

This is necessary as some some data needs to be stored into the class'
hidden property `__vccOpts`, which might not yet be available at the
initialisation and configuration phase of the builder. As soon as all
Vue properties, methods and whatsoever have been configured with the
builder, it is possible to create the `__vccOpts` on the class and
then create an instance. This is done in the final phase as soon as
the builder actually builds-up the component's class.

BREAKING CHANGE: init builder with instance is optional

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>