refactor(setup): `setup()` creates rendering context
Unfortunately, Vue's render function does not cope well with a
rendering context that is an instance of a class. It seems as the
context (a.k.a. "Public Instance") was never meant to be an instance
of a class but only a plain object. It has at least these issues:

* not setting any `this` context for member functions
* complaining about properties with prefix "_" although this is
  widely used with private members of classes
* ignore any inherited property and only access "own" properties

Given the attitude to the TypeScript support of Vue, this seems to be
a deliberate choice. However, to make that work the `setup()` factory
function returns a wrapping proxy to the Vue library that intercepts
access to the properties and fix these issues.

This refactoring will make the `setup()` function to create the
rendering context in order to make the component builder stick with
creating the component instance and not bothering with template
issues.

Extracting the function to a separate file reduces the code of the
component builder making it easier to read.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>