fix(builder): block re-use of same builder
The component builder manipulates the instance on-the-fly and does
not store all alteration to apply them later. Hence, new instances
of components can not be prepared in-a-row. Instead, the whole
procedure must be performed again. Hence, it is the best to block
the re-use of the builder by blocking re-assigning a different
instance to manipulate.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>