fix(utils): detecting inherited props ignore `Object`
The function to detect all inherited properties failed to exclude the
base class `Object` for plain JavaScript objects. Since these are not
instances of a class, the do not inherit any properties from any class
and this should just return an empty list.
This case was missing and not handled before.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>