fix(decorator): fix detection of getter/setter computed
The code used `return` instead of `continue` to skip properties of
no-interest. It seems as if the code previously used a `forEach`
callback function on the array instead of the `for`-loop.
Hence, detection of getters and setters always returned `undefined` as
each component class contains at least one property to skip
(namely `constructor`).

Previous refactoring has failed and is fixed now.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>