fix(builder): enable overwrite of defined props
Properties defined by the builder are set to be redefined again later.
The builder might be used at different stages, each stage defining
an overlapping set of properties. The later definition replaces the
former.

This is necessary as getters/setters (computed values) might be
overloaded in child classes. So, the parent setup function defines a
property for the computed value and afterwards, the child setup does
the same with a different value. This is perfectly fine and intended.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>