fix(builder): make value-properties responsive
Value based properties defined on a class are part of the "data"-set
for a Vue component. As such, they need to be responsive, too.
This commit converts these properties to responsive references to
make Vue react on changes to these values.

This does not apply to getters or setters as these are handled as
"computed" values.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>