test(render): check SFC context returns unique keys
A proxy is returned as render context in order to live patch the
return values as Vue does not handle them correctly:

* bind `this` context to function values of properties - enabling to
  call class member functions.
* hide "private" properties prefixed with "_" or "$" as Vue complains
  about them and may block these in the future
* present any inherited property as "own" property because Vue does
  not check for inherited properties

The list of "own" properties need to be unique. This is checked for
SFC components now. Previous commit has added a Vue component to be
used for that.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>