refactor(release)!: rename package to vue-oop-component
In order to enable Vue components to work on Vue 2 and Vue 3 almost
seamlessly, a special package for Vue 2 is planned. Using the very
same name for the Vue 2 and Vue 3 package enables any component to
work with any Vue version without changes to the imports.

The Vue 2 will use version numbers 2.x.x and Vue 3 package will use
3.x.x. Then the packages will share the major version with the Vue
package they depend on.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>