refactor(release)!: rename package distribution file
As it is planned to support ESM and CommonJS in future releases, the
name of the package distribution file is renamed to contain `.es` as
extension prefix. Additionally, it seems a good idea to rename the
file to a more verbose file name. This will be of help in case the
distribution file is loaded by any browser loader from a common
storage. In such cases, it is very annoying if multiple package named
their distribution file `index`.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>