refactor(vue2): use TypeScript to create transition
The transition file is created using TypeScript. This is necessary as
Vite does not create a separate file for the imported "vue/index.ts"
file. This makes selective import of just the vue part in the
transition file impossible.

Only if Vite knows about this second file entry is it possible to
make Vite provide the necessary files for import into the
`vue2-transition` file.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>