refactor: rename `index.ts` to package name
By using the package name as file name (`nexxar-vue-oop-component.ts`)
the TypeScript compiler emits type declaration in a file of the same
name. Now, the package distribution file has the same base name as
its type declaration file.

There might be problems with importing the library and typeScript
if the files use a different base name. At least that is a suggestion
I have read various times in blog posts and forum posts.

However, whether this issue exists or not, it does not hurt to rename
the source file, too.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>